# server/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"_36_Enums.Crop": {
				"type": "string",
				"enum": [
					"MAIZ",
					"TRIGO",
					"SORGO",
					"CAFE",
					"AGAVE"
				]
			},
			"_36_Enums.MxState": {
				"type": "string",
				"enum": [
					"JALISCO",
					"MICHOACAN",
					"SINALOA",
					"CHIHUAHUA",
					"SONORA"
				]
			},
			"Quote": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clientName": {
						"type": "string"
					},
					"crop": {
						"$ref": "#/components/schemas/_36_Enums.Crop"
					},
					"state": {
						"$ref": "#/components/schemas/_36_Enums.MxState"
					},
					"areaHa": {
						"type": "number",
						"format": "double"
					},
					"insuredAmount": {
						"type": "number",
						"format": "double"
					},
					"validFrom": {
						"type": "string",
						"format": "date-time"
					},
					"validTo": {
						"type": "string",
						"format": "date-time"
					},
					"polygon": {},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"clientName",
					"crop",
					"state",
					"areaHa",
					"insuredAmount",
					"validFrom",
					"validTo",
					"polygon",
					"ownerId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ValidationErrorDetails": {
				"properties": {},
				"type": "object",
				"additionalProperties": {
					"properties": {
						"message": {}
					},
					"required": [
						"message"
					],
					"type": "object"
				}
			},
			"Record_string.any_": {
				"properties": {},
				"additionalProperties": {},
				"type": "object",
				"description": "Construct a type with a set of properties K of type T"
			},
			"ValidateError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string",
						"default": ""
					},
					"stack": {
						"type": "string"
					},
					"statusCode": {
						"type": "number",
						"format": "double",
						"default": 400
					},
					"details": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/ValidationErrorDetails"
							},
							{
								"$ref": "#/components/schemas/Record_string.any_"
							}
						]
					}
				},
				"required": [
					"name",
					"message",
					"statusCode"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IQuote": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clientName": {
						"type": "string"
					},
					"crop": {
						"$ref": "#/components/schemas/_36_Enums.Crop"
					},
					"state": {
						"$ref": "#/components/schemas/_36_Enums.MxState"
					},
					"areaHa": {
						"type": "number",
						"format": "double"
					},
					"insuredAmount": {
						"type": "number",
						"format": "double"
					},
					"validFrom": {
						"type": "string",
						"format": "date-time"
					},
					"validTo": {
						"type": "string",
						"format": "date-time"
					},
					"polygon": {},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"clientName",
					"crop",
					"state",
					"areaHa",
					"insuredAmount",
					"validFrom",
					"validTo",
					"polygon",
					"ownerId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_IQuote_": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"clientName": {
						"type": "string"
					},
					"crop": {
						"$ref": "#/components/schemas/_36_Enums.Crop"
					},
					"state": {
						"$ref": "#/components/schemas/_36_Enums.MxState"
					},
					"areaHa": {
						"type": "number",
						"format": "double"
					},
					"insuredAmount": {
						"type": "number",
						"format": "double"
					},
					"validFrom": {
						"type": "string",
						"format": "date-time"
					},
					"validTo": {
						"type": "string",
						"format": "date-time"
					},
					"polygon": {},
					"ownerId": {
						"type": "number",
						"format": "double"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
			"NodeJS.ReadableStream": {
				"properties": {
					"readable": {
						"type": "boolean"
					}
				},
				"required": [
					"readable"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "server",
		"version": "1.0.0",
		"description": "Sinecta Quotes",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "victor.barreratoledo@gmail.com"
		}
	},
	"paths": {
		"/api/v1/quote": {
			"post": {
				"operationId": "Create",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Quote"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unauthorized",
											"message": "Validation failed",
											"statusCode": 401,
											"details": {}
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Failed",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"message": "Validation failed",
											"name": "Quote Schema Validation Failed",
											"statusCode": 422,
											"details": {
												"body.$0.name": {
													"message": "'clientName' is required"
												}
											}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Post",
					"Quotes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IQuote"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "List",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unauthorized",
											"message": "Validation failed",
											"statusCode": 401,
											"details": {}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Get",
					"Quotes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "crop",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "q",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/quote/{id}": {
			"put": {
				"operationId": "Update",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unauthorized",
											"message": "Validation failed",
											"statusCode": 401,
											"details": {}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Quotes"
				],
				"security": [
					{
						"jwt": [
							"admin"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_IQuote_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "Get",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"400": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"error": {
											"type": "string"
										}
									},
									"required": [
										"error"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unauthorized",
											"message": "Validation failed",
											"statusCode": 401,
											"details": {}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Get",
					"Quotes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ValidateError"
								},
								"examples": {
									"Example 1": {
										"value": {
											"name": "Unauthorized",
											"message": "Validation failed",
											"statusCode": 401,
											"details": {}
										}
									}
								}
							}
						}
					}
				},
				"tags": [
					"Quotes"
				],
				"security": [
					{
						"jwt": [
							"ADMIN"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/v1/quote/export/csv": {
			"get": {
				"operationId": "ExportCSV",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Quotes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/quote/export/pdf": {
			"get": {
				"operationId": "ExportPDF",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/NodeJS.ReadableStream"
								}
							}
						}
					}
				},
				"tags": [
					"Quotes"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/api/v1/auth/signin": {
			"post": {
				"operationId": "Signin",
				"responses": {
					"200": {
						"description": "OK",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {
											"properties": {
												"refreshToken": {
													"type": "string"
												},
												"accessToken": {
													"type": "string"
												}
											},
											"required": [
												"refreshToken",
												"accessToken"
											],
											"type": "object"
										}
									},
									"required": [
										"data"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Signin",
					"Auth"
				],
				"security": [],
				"parameters": [
					{
						"in": "header",
						"name": "authorization",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}