# server/openapi-client/models/ValidationErrorDetails.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ValidationErrorDetails = Record<string, {
    message: any;
}>;